+++ src/rpc/mod.rs
@@ -87,9 +87,12 @@ macro_rules! for_each_method {
         $callback!(crate::rpc::eth::EthGetTransactionByHash);
         $callback!(crate::rpc::eth::EthGetTransactionCount);
         $callback!(crate::rpc::eth::EthGetTransactionHashByCid);
+        $callback!(crate::rpc::eth::EthGetTransactionByBlockNumberAndIndex);
+        $callback!(crate::rpc::eth::EthGetTransactionByBlockHashAndIndex);
         $callback!(crate::rpc::eth::EthMaxPriorityFeePerGas);
         $callback!(crate::rpc::eth::EthProtocolVersion);
         $callback!(crate::rpc::eth::EthNewFilter);
+        $callback!(crate::rpc::eth::EthNewBlockFilter);
         $callback!(crate::rpc::eth::EthSyncing);
         $callback!(crate::rpc::eth::Web3ClientVersion);